requirements.txt
pyyaml
fpdf
pandas
odfpy
xlrd
